Reorganize plus components into a plus folder
