+++ clients/admin-ui/src/features/dataset/DatasetHeading.tsx
@@ -3,7 +3,7 @@ import { useMemo } from "react";
 import { useSelector } from "react-redux";
 
 import { StepperCircleCheckmarkIcon } from "~/features/common/Icon";
-import { selectActiveClassifyDataset } from "~/features/common/plus.slice";
+import { selectActiveClassifyDataset } from "~/features/plus/plus.slice";
 import { ClassificationStatus } from "~/types/api";
 
 const DatasetHeading = () => {
